Remove unnecessary connection acquire and release dance
The connection pool already handles this stuff for me. So it's better
to just open a "new" connection per query. It should be noted that
these connections are *already* open and I'm just cycling through them.

This also partially improves the type hinting situation to allow better
manual narrowing with asyncpg.Record objects